fixed(?) doxygen warning
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@5989 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/sound/WavWriter.cc
@@ -77,7 +77,7 @@ void WavWriter::write8mono(unsigned char val)
 	write8mono(&val, 1);
 }
 
-void WavWriter::write8mono(unsigned char* val, size_t len)
+void WavWriter::write8mono(unsigned char* val, unsigned len)
 {
 	bytes += fwrite(val, 1, len, wavfp);
 }
